% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SANETPA-package.R
\docType{package}
\name{SANETPA-package}
\alias{SANETPA}
\alias{SANETPA-package}
\title{SANETPA: Research Compendium for a Study of Sexual Assault Nurse Examiner Training Program Attrition}
\description{
This package is a research compendium for a study examining attrition from a sexual assault nurse examininer (SANE) training program. The research team is led by Dr. Katherine Dontje (PI) and Dr. Rebecca Campbell (Co-I) at Michigan State University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjpierce/SANETPA}
}

}
\author{
\strong{Maintainer}: Steven J. Pierce \email{pierces1@msu.edu} (\href{https://orcid.org/0000-0002-0679-3019}{ORCID}) [copyright holder]

}
\keyword{internal}
